requirements.txt
google-cloud-vision
pillow
flask
opencv-python
